add getFieldValue
